feat : Clinic 페이지 클라이언트 페이지로 변경 (#96)
