+++ src/features/clinic/ScheduleSection.tsx
@@ -1,5 +1,7 @@
 'use client'
 
+import { useRouter } from 'next/navigation'
+
 import { Label } from '@/components'
 import ActionTag from '@/components/actionTag/ActionTag'
 import { useCurrentDate } from '@/store'
@@ -10,14 +12,19 @@ import { ScheduleItem } from './ScheduleItem'
 type ClinicListProps = ClinicResponse
 
 export const ScheduleSection = ({ list }: ClinicListProps) => {
+  const router = useRouter()
   const date = useCurrentDate()
   const todaySchedules = list.filter((item) => item.hospitalDate.startsWith(date))
+
+  const handleClickPlusButton = () => {
+    router.push('clinic/write')
+  }
 
   return (
     <section className="mt-7">
       <div className="flex-between mb-3">
         <Label icon="clinic-label">병원 내원 일정</Label>
-        <ActionTag.Plus label="추가" primary onClick={() => {}} />
+        <ActionTag.Plus label="추가" primary onClick={handleClickPlusButton} />
       </div>
 
       <div className="flex-column gap-3">
