+++ src/features/clinic/index.ts
@@ -1,3 +1,4 @@
+export * from './ClinicClientPage'
 export * from './FloatingProfile'
 export * from './ScheduleItem'
 export * from './ScheduleSection'
